コメント対応
